fix thumbnail circular dependencies
