+++ libs/platform/src/lib/components/thumbnail/public_api.ts
@@ -1,4 +1,3 @@
 export * from './thumbnail.module';
 export * from './thumbnail.component';
-
-
+export * from './thumbnail.interfaces';
